test to solve artifacts : cube

+++ models/mesh_texturize.py
@@ -44,7 +44,7 @@ class TexturizeModel:
 
     def set_input(self, data):
         input_edge_features = torch.from_numpy(data['edge_features']).float()
-        labels = torch.from_numpy(data['label']).float()
+        labels = torch.from_numpy(data['label']).int()
         # set inputs
         self.edge_features = input_edge_features.to(self.device).requires_grad_(self.is_train)
         self.labels = labels.to(self.device)
